Fixes for spawning

+++ Assets/Code/NPC.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class NPC : MonoBehaviour {
 
-	public int omega;
+	public float omega;
 	// Use this for initialization
 	void Start () {
